+++ docs/examples/timepicker/ExSimple.vue
@@ -1,0 +1,8 @@
+<template>
+    <b-field label="Select time">
+        <b-timepicker
+            placeholder="Click to select..."
+            icon="access_time">
+        </b-timepicker>
+    </b-field>
+</template>