+++ src/utils/FormElementMixin.js
@@ -77,7 +77,7 @@ export default {
         checkHtml5Validity() {
             if (this.$refs[this.$data._elementRef] === undefined) return
 
-            const el = this.$data._isAutocomplete || this.$data._isDatepicker
+            const el = this.$data._isAutocomplete || this.$data._isDatepicker || this.$data._isTimepicker
                 ? this.$refs.input.$refs.input
                 : this.$refs[this.$data._elementRef]