+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Breaking changes
 
+* **New component: Timepicker**, [see docs](https://buefy.github.io/#/documentation/timepicker)
 * Vue 2.5+ is now the minimum required version
 * **Removed** prompt dialog ``inputPlaceholder``, ``inputName`` and ``inputMaxlength`` props
 * Add ``disabled`` prop to tabItem